requirements.txt
numpy==1.21.0
opencv-python==4.5.3.56
pytesseract==0.3.8
Pillow==8.3.1
matplotlib==3.4.3
